+++ src/links.rs
@@ -22,4 +22,5 @@ pub mod users {
     pub const UNFOLLOW: &'static str = "https://api.twitter.com/1.1/friendships/destroy.json";
     pub const FRIENDSHIPS_INCOMING: &'static str = "https://api.twitter.com/1.1/friendships/incoming.json";
     pub const FRIENDSHIPS_OUTGOING: &'static str = "https://api.twitter.com/1.1/friendships/outgoing.json";
+    pub const FRIENDSHIP_SHOW: &'static str = "https://api.twitter.com/1.1/friendships/show.json";
 }