+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - New functions to follow/unfollow a user (`follow`, `unfollow`)
 - New function to test the validity of access tokens (`verify_tokens`)
 - New functions to see incoming/outgoing follow requests (`incoming_requests`, `outgoing_requests`)
+- New function to see friendship status between users (`relation`)
 
 ### Changed
 - Combined IDLoader and UserLoader into CursorIter (with the same interface)